Update to v098r16 release.
byuu says:

Changelog:
- GNUmakefile: reverted $(call unique,) to $(strip)
- processor/r6502: removed templates; reduces object size from 146.5kb
  to 107.6kb
- processor/lr35902: removed templates; reduces object size from 386.2kb
  to 197.4kb
- processor/spc700: merged op macros for switch table declarations
- sfc/coprocessor/sa1: partial cleanups; flattened directory structure
- sfc/coprocessor/superfx: partial cleanups; flattened directory structure
- sfc/coprocessor/icd2: flattened directory structure
- gb/ppu: changed behavior of STAT IRQs

Major caveat! The GB/GBC STAT IRQ changes has a major bug in it somewhere
that's seriously breaking most games. I'm pushing the WIP anyway, because
I believe the changes to be mostly correct. I'd like to get more people
looking at these changes, and also try more heavy-handed hacking and
diff comparison logging between the previous WIP and this one.

+++ higan/processor/spc700/registers.hpp
@@ -4,15 +4,15 @@ struct Flag {
          | (h << 3) | (i << 2) | (z << 1) | (c << 0);
   }
 
-  inline auto operator=(uint8 data) -> uint {
+  inline auto operator=(uint data) -> uint {
     n = data & 0x80; v = data & 0x40; p = data & 0x20; b = data & 0x10;
     h = data & 0x08; i = data & 0x04; z = data & 0x02; c = data & 0x01;
     return data;
   }
 
-  inline auto operator|=(uint8 data) -> uint { return operator=(operator uint() | data); }
-  inline auto operator^=(uint8 data) -> uint { return operator=(operator uint() ^ data); }
-  inline auto operator&=(uint8 data) -> uint { return operator=(operator uint() & data); }
+  inline auto operator|=(uint data) -> uint { return operator=(operator uint() | data); }
+  inline auto operator^=(uint data) -> uint { return operator=(operator uint() ^ data); }
+  inline auto operator&=(uint data) -> uint { return operator=(operator uint() & data); }
 
   bool n, v, p, b, h, i, z, c;
 };
@@ -24,8 +24,8 @@ struct Word {
   inline auto operator++() -> uint { return ++w; }
   inline auto operator--() -> uint { return --w; }
 
-  inline auto operator++(int) -> uint { unsigned data = w++; return data; }
-  inline auto operator--(int) -> uint { unsigned data = w--; return data; }
+  inline auto operator++(int) -> uint { uint data = w++; return data; }
+  inline auto operator--(int) -> uint { uint data = w--; return data; }
 
   inline auto operator+=(uint data) -> uint { return w += data;; }
   inline auto operator-=(uint data) -> uint { return w -= data;; }
